+++ src/DataService/Media/Media.php
@@ -173,7 +173,7 @@ class Media extends ObjectBase implements PublicIdWithGuidInterface
     /**
      * The streamingUrl of the default thumbnail for this Media.
      *
-     * @var string
+     * @var Uri
      */
     protected $defaultThumbnailUrl;
 
@@ -813,12 +813,12 @@ class Media extends ObjectBase implements PublicIdWithGuidInterface
      * Returns the streamingUrl of the default thumbnail for this Media.
      *
      * Note that the defaultThumbnailUrl may include dynamic URL substitution
-     * parameters. Use getDefaultThumbnailUri() to get a Uri object with
-     * dynamic URL substitution parameters resolved.
+     * parameters. Use getNormalizedDefaultThumbnailUrl() to get a Uri object
+     * with dynamic URL substitution parameters resolved.
      *
      * @see https://docs.theplatform.com/help/media-delivery-urls
      */
-    public function getDefaultThumbnailUrl(): string
+    public function getDefaultThumbnailUrl(): Uri
     {
         return $this->defaultThumbnailUrl;
     }
@@ -826,7 +826,7 @@ class Media extends ObjectBase implements PublicIdWithGuidInterface
     /**
      * Set the streamingUrl of the default thumbnail for this Media.
      *
-     * @param string
+     * @param Uri
      */
     public function setDefaultThumbnailUrl($defaultThumbnailUrl)
     {
@@ -834,7 +834,7 @@ class Media extends ObjectBase implements PublicIdWithGuidInterface
     }
 
     /**
-     * Get the default thumbnail uri with dynamic URL substitution applied.
+     * Get the default thumbnail URL with dynamic URL substitution applied.
      *
      * Note that the defaultThumbnailUrl may include dynamic URL substitution
      * parameters. Here we make an attempt at substituting known parameters.
@@ -844,14 +844,14 @@ class Media extends ObjectBase implements PublicIdWithGuidInterface
      *
      * @param bool $ssl TRUE for SSL substitution, otherwise FALSE.
      *
-     * @return \GuzzleHttp\Psr7\Uri Uri object for the default thumbnail url
+     * @return \GuzzleHttp\Psr7\Uri A Uri object for the default thumbnail url
      *                              with known dynamic URL substitution done.
      *
      * @see https://docs.theplatform.com/help/media-delivery-urls#tp-toc38
      */
-    public function getDefaultThumbnailUri(bool $ssl = true): Uri
+    public function getNormalizedDefaultThumbnailUrl(bool $ssl = true): Uri
     {
-        $defaultThumbnailUrl = urldecode($this->getDefaultThumbnailUrl());
+        $defaultThumbnailUrl = urldecode((string) $this->getDefaultThumbnailUrl());
         $defaultThumbnailUrl = preg_replace('~\{ssl:(.*?):(.*?)\}~', $ssl ? '$1' : '$2', $defaultThumbnailUrl);
 
         return new Uri($defaultThumbnailUrl);